pyproject.toml
[project]
name = "ocr-ai"
version = "0.1.0"
description = "DeepSeek-OCR on Apple Silicon"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.11.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "torch>=2.9.1",
    "transformers>=4.46.0",
    "accelerate>=0.34.0",
    "sentencepiece>=0.2.0",
    "huggingface-hub>=0.26.0",
    "ollama>=0.6.1",
    "dotenv>=0.9.9",
]
